renamed hands to hand
fixes #46
