+++ tests/hand.js
@@ -3,7 +3,7 @@
 const expect = require("chai").expect;
 
 const deckModel = require("../models/deck");
-const handsModel = require("../models/hands");
+const handModel = require("../models/hand");
 
 let deck;
 let newDeck;
@@ -13,7 +13,7 @@ describe("Hands Model - New Hands", () => {
 	before(() => {
 		deck = deckModel.newDeck(52);
 		newDeck = deckModel.shuffleDeck(deck);
-		hands = handsModel.newHands(newDeck);
+		hands = handModel.newHands(newDeck);
 	});
 
 	it("hands should be a valid array", (done) => {
