+++ tests/player.js
@@ -4,7 +4,7 @@ const expect = require("chai").expect;
 
 const playerModel = require("../models/player");
 const deckModel = require("../models/deck");
-const handModel = require("../models/hands");
+const handModel = require("../models/hand");
 
 let player;
 let hands;
